requirements.txt
discord.py
discord-py-slash-command
wavelink